general maintenance

+++ Future_Dev/Controllers/CoursesController.cs
@@ -8,7 +8,7 @@ namespace Future_Dev.Controllers
 
     [ApiController]
     [Route("api/courses")]
-    [Authorize(Roles ="Admin")]
+    [Authorize]
     public class CoursesController : ControllerBase
     {
         private readonly ICourseService _courseService;
@@ -26,6 +26,7 @@ namespace Future_Dev.Controllers
         }
 
         [HttpGet("GetCourse/{id}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> GetById(int id)
         {
             var course = await _courseService.GetCourseById(id);
@@ -36,6 +37,7 @@ namespace Future_Dev.Controllers
         }
 
         [HttpPost("CreateCourse")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Create([FromBody] CreateCourseDto dto)
         {
             await _courseService.CreateCourse(dto);
@@ -43,6 +45,7 @@ namespace Future_Dev.Controllers
         }
 
         [HttpPut("UpdateCourse/{id}")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateCourseDto dto)
         {
             await _courseService.UpdateCourse(id, dto);
@@ -52,33 +55,39 @@ namespace Future_Dev.Controllers
         [HttpDelete("DeleteCourse/{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!User.IsInRole("Admin"))
+                return Forbid("Only Admins can delete students");
+
             await _courseService.DeleteCourse(id);
             return Ok("Course deleted successfully.");
         }
 
         [HttpGet("SearchCourse")]
+        [Authorize(Roles = "Admin,User")]
         public async Task<IActionResult> Search([FromQuery] string title)
         {
             return Ok(await _courseService.SearchCourses(title));
         }
 
-        [Authorize(Roles = "User,Admin")]
+        [Authorize(Roles = "User")]
         [HttpGet("MyCourses")]
         public async Task<IActionResult> GetMyCourses()
         {
             return Ok(await _courseService.GetMyCourses());
         }
 
-        [HttpGet("GetCourse")]
+        [HttpGet("GetCourseByUser")]
+        [Authorize(Roles = "Admin")]
         public async Task<IActionResult> GetCoursesByStudentId([FromQuery] int studentId)
         {
             return Ok(await _courseService.GetCoursesByStudentId(studentId));
         }
 
-        [HttpGet("GetCourseByUser")]
-        public async Task<IActionResult> GetCoursesByUserId([FromQuery] int studentId)
+        [HttpGet("GetUsersByCourse")]
+        [Authorize(Roles = "Admin")]
+        public async Task<IActionResult> GetStudentsByCourseId([FromQuery] int courseId)
         {
-            return Ok(await _courseService.GetMyCourses());
+            return Ok(await _courseService.GetStudentsByCourseId(courseId));
         }
 
     }
